mc/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-admin-interface==0.22.1
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==3.0.1
django-jazzmin==2.6.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.6
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
